separate constraints (dynamics)
